keshavoct98/requirements.txt
absl-py==0.9.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
easydict==1.9
easyocr==1.1.10
editdistance==0.5.3
efficientnet==1.0.0
essential-generators==0.9.2
fonttools==4.16.1
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==2.0.0
Keras-Applications==1.0.8
keras-ocr==0.8.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.2
networkx==2.5
numpy==1.19.3
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.3.0
pandas==1.1.4
Pillow==8.0.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.2.0
pyparsing==2.4.7
pytesseract==0.3.6
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.4.1
Shapely==1.7.1
six==1.15.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
tifffile==2020.10.1
torch==1.7.0+cu101
torchaudio==0.7.0
torchvision==0.8.1+cu101
tqdm==4.51.0
typing-extensions==3.7.4.3
urllib3==1.25.11
validators==0.18.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
